Add optional margins to limit follow area
Useful when recording only part of the screen on huge monitors.
E.g.: on a 32:9 monitor with resolution 5120x1440, if recording only the right side, of 2560x1440, set "Zoom left margin" to 2560. When zoomed in, the follow window will stop following the mouse when it goes left of the middle of the screen.

This commit also renames variables for easier code readability.